src/unlam/paradigmas/objetos/ej01/Nota.java
package unlam.paradigmas.objetos.ej01;

public class Nota {
	private static int notaAprobacion = 7;
	private int nota;
	
    /**
     * pre : valorInicial está comprendido entre 0 y 10.
     * post: inicializa la Nota con el valor indicado.
     */
    public Nota(int valorInicial) {
    	this.nota = valorInicial;
    }
    
    /**
     * post: devuelve el valor numérico de la Nota,
     *       comprendido entre 0 y 10.
     */
    public int obtenerValor() {
    	return this.nota;
    }
    
    /**
     * post: indica si la Nota permite o no la aprobación.
     */
    public boolean aprobado() {
    	return this.nota >= notaAprobacion;
    }
    
    /**
     * post: indica si la Nota implica desaprobación.
     */
    public boolean desaprobado() {
    	return this.nota < notaAprobacion;
    }
}
